impl: recursive curriculum structure object

+++ internal/core/model/sys_category_course_requisite.go
@@ -11,7 +11,7 @@ const (
 )
 
 type SysCategoryCourseRequisite struct {
-	ID                int64         `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID                int           `gorm:"primaryKey;autoIncrement" json:"id"`
 	CategoryCoursesID int           `gorm:"column:category_courses_id;type:int;not null" json:"category_courses_id"`
 	RelatedCourseNo   string        `gorm:"column:related_course_no;type:varchar(6);not null" json:"related_course_no"`
 	RequisiteType     RequisiteType `gorm:"column:requisite_type;type:requisite_type;not null" json:"requisite_type"`
